refactor(materials): build for materials, move files
