+++ packages/frameworks/vue/src/chat/GenuiChat.vue
@@ -19,7 +19,6 @@ import type {
 import { ref, watch, computed, h, inject, nextTick } from 'vue';
 import type { Ref } from 'vue';
 import { CustomModelProvider } from './CustomModelProvider';
-import SchemaRenderer from '../renderer/SchemaCardRenderer.vue';
 import { scrollEnd, throttle, toSlotFunction } from './chat-utils';
 import { useFileUpload } from './useFileUpload';
 import AttachmentsRenderer from './renderer/AttachmentsRenderer.vue';
@@ -33,6 +32,7 @@ import GeneratingComponent from './GeneratingComponent.vue';
 import { useContinueChatAction } from './continue-chat-action';
 import type { IMessageItem } from '@opentiny/genui-sdk-core';
 import type { IRendererProps } from '../renderer';
+import { SchemaRenderer } from '../renderer';
 import ErrorText from './ErrorText.vue';
 import { useResize } from './composable/use-resize';
 import { useConversation } from './tiny-robot-patch/useConversation';
