+++ packages/materials/angular-opentiny-ng/src/index.ts
@@ -1,0 +1,1 @@
+export * from './render-config';
