+++ packages/materials/angular-opentiny-ng/src/render-config/example-schema.ts
