+++ packages/materials/angular-opentiny-ng/src/render-config/index.ts
