+++ packages/materials/angular-opentiny-ng/src/render-config/merge.ts
