+++ packages/materials/angular-opentiny-ng/src/render-config/white-list.ts
