+++ packages/materials/angular-opentiny-ng/vite.config.ts
@@ -1,0 +1,26 @@
+import path from 'path';
+import { defineConfig } from 'vite';
+import dts from 'vite-plugin-dts';
+import escapeStringRegexp from 'escape-string-regexp';
+import packageJson from './package.json';
+
+export default defineConfig({
+  plugins: [
+    dts({
+      rollupTypes: true,
+    }),
+  ],
+  build: {
+    lib: {
+      entry: {
+        index: path.resolve(__dirname, './src/index.ts'),
+        'render-config': path.resolve(__dirname, './src/render-config/index.ts'),
+      },
+      formats: ['es'],
+      fileName: (format, entryName) => `${entryName}.js`, //TODO: CSS 打包不会自动引入
+    },
+    rollupOptions: {
+      external: [...Object.keys(packageJson.dependencies || {}).map(name => new RegExp(`^${escapeStringRegexp(name)}(/|$)`))],
+    },
+  },
+});
