+++ packages/materials/vue-opentiny-vue/package.json
@@ -1,14 +1,46 @@
 {
-    "name": "@opentiny/genui-sdk-materials-vue-opentiny-vue",
-    "version": "0.0.1",
-    "main": "dist/index.js",
-    "types": "dist/index.d.ts",
-    "type": "module",
-    "scripts": {
-        "build": "vite build"
+  "name": "@opentiny/genui-sdk-materials-vue-opentiny-vue",
+  "version": "0.0.1",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "type": "module",
+  "scripts": {
+    "build": "vite build"
+  },
+  "dependencies": {
+    "vue": "^3.5.13",
+    "@opentiny/vue-button": "^3.25.0",
+    "@opentiny/vue-config-provider": "^3.25.0",
+    "@opentiny/vue-chart-bar": "~3.14.0",
+    "@opentiny/vue-chart-histogram": "~3.14.0",
+    "@opentiny/vue-chart-line": "~3.14.0",
+    "@opentiny/vue-chart-pie": "~3.14.0",
+    "@opentiny/vue-chart-radar": "~3.14.0",
+    "@opentiny/vue-chart-ring": "3.14.0",
+    "@opentiny/vue-select": "^3.25.0"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^5.0.0",
+    "vite": "^6.0.0",
+    "vite-plugin-dts": "^4.0.0",
+    "escape-string-regexp": "^1.0.5",
+    "typescript": "^5.8.2"
+  },
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.js"
     },
-    "dependencies": {
-        "vue": "^3.5.13"
+    "./render-config": {
+      "import": "./dist/render-config.js",
+      "types": "./dist/render-config.d.ts",
+      "require": "./dist/render-config.js"
+    },
+    "./extend-renderer": {
+      "import": "./dist/extend-renderer.js",
+      "types": "./dist/extend-renderer.d.ts",
+      "require": "./dist/extend-renderer.js"
     }
-
+  }
 }
