+++ packages/materials/vue-opentiny-vue/src/extend-renderer.ts
@@ -4,7 +4,11 @@ import ActionButton from './components/ActionButton.vue';
 import TinyTabsWrap from './components/TinyTabsWrap.vue';
 import GridStack from './components/GridStack.vue';
 import GridStackItem from './components/GridStackItem.vue';
+
+export let extened = false;
 export const exentdMapper = (Mapper: any, customComponents: Record<string, Component>) => {
+  if (extened) return;
+  extened = true;
   Mapper.TinySelect = TinySelectWrap;
   Mapper.ActionButton = ActionButton;
   Mapper.TinyTabs = TinyTabsWrap;
