+++ packages/materials/vue-opentiny-vue/src/index.ts
@@ -1,0 +1,2 @@
+export * from './render-config';
+export * from './extend-renderer';
