+++ packages/materials/vue-opentiny-vue/src/render-config/chart.json
