+++ packages/materials/vue-opentiny-vue/src/render-config/example-schema.ts
