+++ packages/materials/vue-opentiny-vue/src/render-config/extend.json
