+++ packages/materials/vue-opentiny-vue/src/render-config/index.ts
