+++ packages/materials/vue-opentiny-vue/src/render-config/merge.ts
