+++ sites/playground/server/package.json
@@ -15,9 +15,10 @@
     "@ai-sdk/deepseek": "^1.0.5",
     "@ai-sdk/openai": "^2.0.10",
     "@modelcontextprotocol/sdk": "^1.13.0",
-    "@modelcontextprotocol/server-sequential-thinking": "^2025.7.1",
     "@openrouter/ai-sdk-provider": "^1.1.2",
     "@opentiny/genui-sdk-core": "workspace:*",
+    "@opentiny/genui-sdk-materials-vue-opentiny-vue": "workspace:*",
+    "@opentiny/genui-sdk-materials-angular-opentiny-ng": "workspace:*",
     "ai": "5.0.10",
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
