+++ sites/playground/server/tsconfig.dev.json
@@ -5,6 +5,12 @@
     "paths": {
       "@opentiny/genui-sdk-core": [
         "../../../packages/core/src/index.ts"
+      ],
+      "@opentiny/genui-sdk-materials-vue-opentiny-vue/render-config": [
+        "../../../packages/materials/vue-opentiny-vue/src/render-config/index.ts"
+      ],
+      "@opentiny/genui-sdk-materials-angular-opentiny-ng/render-config": [
+        "../../../packages/materials/angular-opentiny-ng/src/render-config/index.ts"
       ]
     }
   },
